+++ ChangeLog.md
@@ -23,6 +23,7 @@ Distributions that do not include "beta" in the tag name correspond to the major
 announced releases. The source distributions associated with all releases, major
 or beta, are equally accessible as tarballs through the Github interface.
 
+- Version 5.7.9beta47    November 15, 2022
 - Version 5.7.9beta46    November 11, 2022
 - Version 5.7.9beta45    November 6, 2022
 - Version 5.7.9beta44    August 9, 2022
@@ -422,6 +423,13 @@ or beta, are equally accessible as tarballs through the Github interface.
 --
 ### MB-System Version 5.7 Release Notes:
 --
+
+#### 5.7.9beta47 (November 15, 2022)
+
+Mbnavadjust: Added ability to import and use multiple reference grids for picking
+global ties of individual swath data sections.
+
+Mbnavadjustmerge: Added options set the mode (XYZ vs XY vs Z) of all global ties.ll
 
 #### 5.7.9beta46 (November 11, 2022)